requirements/base.txt
amqp==1.4.9
anyjson==0.3.3
Babel==2.3.4
billiard==3.3.0.23
celery==3.1.18
defusedxml==0.4.1
Django==1.8
django-crispy-forms==1.6.0
django-haystack==2.4.0
django-taggit==0.17.1
djangorestframework==3.4.5
ecdsa==0.13
Fabric==1.12.0
flower==0.9.1
kombu==3.0.35
Markdown==2.6.2
oauthlib==1.1.2
paramiko==1.17.2
Pillow==2.9.0
psycopg2==2.6.2
pycrypto==2.6.1
PyJWT==1.4.2
pysolr==3.3.2
python-social-auth==0.2.12
python3-openid==3.0.10
pytz==2016.6.1
redis==2.10.3
requests==2.11.0
requests-oauthlib==0.6.2
six==1.10.0
sorl-thumbnail==12.3
sqlparse==0.2.1
tornado==4.2
